Centralized "exec" calls from Java into a couple methods
that are very thorough in logging standard error output, and including
the standard error in exceptions.  This should help us in resolving some
scripts that work from command line, but not when called by the daemon.

Note: The fix to the "mysql_dump_database" script was to define the
"HOME" environment variable before using "--login-path".  HOME was not
defined by the daemon running under systemd.